Basic Game
Game plays fine. Score increments for player wins, decrements for app wins, announces rounds which are ties. Includes a reset button to set the score back to zero. Player can tap on the rock, paper, or scissor emojis to pick their move. Added some gradient styling that looks subpar but wanted to play around with it.